+++ toduck.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "0d7973357ac0370c5d93d71370d3eda9eac99575a9bcc25d475b22a0f416f710",
+  "originHash" : "a0df9cd3e9d0f35a977275373c2326565c9336aaf1c4eb45741e2d925b85bc16",
   "pins" : [
     {
       "identity" : "alamofire",
