+++ toduck/Source/Domain/Entity/PomodoroTimer.swift
@@ -8,12 +8,12 @@
 import Foundation
 import UIKit
 
-struct Time {
+public struct Time: Hashable{
     var hours: Int
     var minutes: Int
     var seconds: Int
     
-    init(hours: Int = 0, minutes: Int = 0, seconds: Int = 0) {
+    public init(hours: Int = 0, minutes: Int = 0, seconds: Int = 0) {
         self.hours = hours
         self.minutes = minutes
         self.seconds = seconds
